+++ first-webapp/src/main/java/com/learnspring/firstwebapp/controllers/CatalogController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import com.learnspring.firstwebapp.entities.Product;
+import com.learnspring.firstwebapp.domain.Product;
 import com.learnspring.firstwebapp.service.ProductService;
 
 @Controller
